config.py
import os

from dotenv import load_dotenv

load_dotenv()

URL_DATABASE = os.getenv("DB_URI")

ACCESS_TOKEN_EXPIRE_MINUTES = int(os.environ.get("ACCESS_TOKEN_EXPIRE_MINUTES"))
SECRET_KEY = os.environ.get("SECRET_KEY")
ALGORITHM = os.environ.get("ALGORITHM")
