requirements.txt
fastapi
psycopg2-binary
uvicorn
sqlalchemy
alembic
python-dotenv
asyncpg
python-multipart
bcrypt==4.0.1
passlib==1.7.4
python-jose